+ Alterado front pra capturar os erros do servidor e mostrar alerts na tela + Alterado front para identificar quando está tentando acessar uma URL "forçada" de um produto ou categoria que não existem + Adicionado validação para dar uma mensagem mais amigável quando extá excluindo uma categoria que tem produtos

+++ src/main/kotlin/com/softfocus/server/api/service/CategoriaService.kt
@@ -1,8 +1,8 @@
 package com.softfocus.server.api.service
 
-import com.softfocus.server.api.util.APIException
 import com.softfocus.server.api.model.Categoria
 import com.softfocus.server.api.repository.CategoriaRepository
+import com.softfocus.server.api.util.APIException
 import org.springframework.stereotype.Service
 import java.util.*
 
@@ -44,6 +44,11 @@ class CategoriaService(private val categoriaRepository: CategoriaRepository) {
         if (!categoriaRepository.existsById(categoriaID)) {
             throw APIException("Categoria não existe : " + categoriaID);
         }
+
+        if (categoriaRepository.findById(categoriaID).get().produtos!!.isNotEmpty()) {
+            throw APIException("Existem produtos vinculados a esta categoria, não será permitido excluir");
+        }
+
         categoriaRepository.deleteById(categoriaID)
     }
 }
